Rename the files app_server for nginx_project: for a good managment create the directory dpl this folder get any functions for this project
