+++ dfc_pkg/main.py
@@ -1,10 +1,10 @@
 #!/usr/bin/python3
 import typer
-import deploy.app_server as app_server
+import deploy.nginx_project as nginx_project
 
 app = typer.Typer()
 
-app.add_typer(app_server.app, name="deploy")
+app.add_typer(nginx_project.app, name="deploy")
 
 if __name__ == "__main__":
     app()
